+++ extension-style/src/main/java/com/mapbox/maps/extension/style/sources/CustomGeometrySource.kt
@@ -1,12 +1,13 @@
 package com.mapbox.maps.extension.style.sources
 
+import com.mapbox.bindgen.Expected
+import com.mapbox.bindgen.None
 import com.mapbox.geojson.Feature
 import com.mapbox.maps.*
-import com.mapbox.maps.extension.style.StyleContract
 import com.mapbox.maps.extension.style.utils.check
 
 /**
- * Custom Vector Source, allows using FeatureCollections.
+ * Custom Geometry Source, allows using FeatureCollections.
  *
  * CustomGeometrySource uses a coalescing model for frequent data updates targeting the same tile id,
  * which means, that the in-progress request as well as the last scheduled request are guaranteed to finish.
@@ -16,13 +17,22 @@ class CustomGeometrySource(
   /**
    * Style source identifier.
    */
-  val id: String,
+  id: String,
   /**
    * Settings for the custom geometry.
    */
-  val options: CustomGeometrySourceOptions
-) : StyleContract.StyleSourceExtension {
-  private var delegate: Style? = null
+  private val options: CustomGeometrySourceOptions,
+) : Source(id) {
+  /**
+   * Get the type of the current source as a String.
+   */
+  override fun getType(): String {
+    return "custom-geometry"
+  }
+
+  override fun addSource(style: Style): Expected<String, None> {
+    return style.addStyleCustomGeometrySource(sourceId, options)
+  }
 
   /**
    * Set tile data of a custom geometry.
@@ -31,7 +41,7 @@ class CustomGeometrySource(
    * @param featureCollection An array with the features to add.
    */
   fun setTileData(tileID: CanonicalTileID, featureCollection: MutableList<Feature>) {
-    delegate?.setStyleCustomGeometrySourceTileData(id, tileID, featureCollection).check()
+    delegate?.setStyleCustomGeometrySourceTileData(sourceId, tileID, featureCollection).check()
   }
 
   /**
@@ -39,8 +49,22 @@ class CustomGeometrySource(
    *
    * @param coordinateBounds Coordinate bounds.
    */
+  @Deprecated(
+    message = "This method is deprecated because of confusing naming.",
+    replaceWith = ReplaceWith("invalidateRegion(coordinateBounds)"),
+    level = DeprecationLevel.WARNING
+  )
   fun invalidRegion(coordinateBounds: CoordinateBounds) {
-    delegate?.invalidateStyleCustomGeometrySourceRegion(id, coordinateBounds).check()
+    invalidateRegion(coordinateBounds)
+  }
+
+  /**
+   * Invalidate region for provided custom geometry source.
+   *
+   * @param coordinateBounds Coordinate bounds.
+   */
+  fun invalidateRegion(coordinateBounds: CoordinateBounds) {
+    delegate?.invalidateStyleCustomGeometrySourceRegion(sourceId, coordinateBounds).check()
   }
 
   /**
@@ -48,18 +72,22 @@ class CustomGeometrySource(
    *
    * @param tileID Identifier of the tile
    */
+  @Deprecated(
+    message = "This method is deprecated because of confusing naming.",
+    replaceWith = ReplaceWith("invalidateTile(tileID)"),
+    level = DeprecationLevel.WARNING
+  )
   fun invalidTile(tileID: CanonicalTileID) {
-    delegate?.invalidateStyleCustomGeometrySourceTile(id, tileID).check()
+    invalidateTile(tileID)
   }
 
   /**
-   * Add the source to the Style.
+   * Invalidate tile for provided custom geometry source.
    *
-   * @param delegate The style delegate
+   * @param tileID Identifier of the tile
    */
-  override fun bindTo(delegate: Style) {
-    this.delegate = delegate
-    delegate.addStyleCustomGeometrySource(id, options).check()
+  fun invalidateTile(tileID: CanonicalTileID) {
+    delegate?.invalidateStyleCustomGeometrySourceTile(sourceId, tileID).check()
   }
 }